requirements.txt
opendatasets
torch==2.0.1
torchvision==0.15.2
torchsummary
pandas
dvc
mlflow==2.2.2
notebook
numpy
matplotlib
seaborn
python-box==6.0.2
pyYAML
pandas
tqdm
ensure==1.0.2
joblib
transformers
datasets
scikit-learn
types-PyYAML
scipy
Flask
Flask-Cors
python-dotenv
-e .